backend/requirements.txt
# Web framework & server
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0

# Database
psycopg2-binary>=2.9.9,<3.0

# AWS
boto3>=1.34.0,<2.0

# Auth & security
bcrypt>=4.1.0,<5.0
PyJWT>=2.8.0,<3.0
email-validator>=2.1.0,<3.0

# MFA (TOTP)
pyotp>=2.9.0,<3.0
qrcode>=7.4.0,<8.0

# HTTP & config
requests>=2.31.0,<3.0
python-multipart>=0.0.6,<1.0
